Refactor testing utilities into a dedicated module
Also, removed dealing with `Tricks` in `new_for_testing`.
This is a preparation for implementing sharing of `AnalysisDatabase` between E2E tests.

commit-id:ab430cf2
